% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRsurvfit.R
\name{plot.DRsurvfit}
\alias{plot.DRsurvfit}
\title{Plot method for DRsurvfit objects}
\usage{
\method{plot}{DRsurvfit}(x, level = c("cluster", "individual"), ...)
}
\arguments{
\item{x}{A \code{DRsurvfit} object.}

\item{level}{\code{"cluster"} or \code{"individual"}.}

\item{...}{Unused.}
}
\value{
The input object \code{x} (invisibly).
}
\description{
Plots survival curves (treatment \code{S1} solid, control \code{S0} dashed)
for either the cluster- or individual-level estimand. If \code{tau} is present
(RMST fit), vertical lines are drawn at \code{tau}.
}
\examples{
\dontrun{
data(dat)
fit_spce <- DRsurvfit(
  dat, survival::Surv(time, event) ~ W1 + W2 + Z1 + Z2 + cluster(M),
  intv = "A", method = "marginal", estimand = "SPCE"
)
plot(fit_spce, level = "cluster")

fit_rmst <- DRsurvfit(
  dat, survival::Surv(time, event) ~ W1 + W2 + Z1 + Z2 + cluster(M),
  intv = "A", method = "frailty", estimand = "RMST", tau = c(0.5,1,2)
)
plot(fit_rmst, level = "individual")
}
}
